Made UI changes
Added python logo to back of cards. Added buttons for stick, twist etc so the game looks better.
There are also buttons for when the player is making a bet. When the player makes a choice, the invalid
choices are covered by a green rectangle. The game also has an icon. This commit includes bug fixes.