Added file i/o
The player can enter a username and if that name is in a file with a
corresponding chips value then the player carries on playing from where
he got up to previously. At the start of each turn his new chip value is written to the file and if no file exists then one is created.